+++ src/lib/waku_store/index.ts
@@ -342,7 +342,7 @@ async function* paginate<T extends Message>(
       if (typeof contentTopic !== "undefined") {
         const decoder = decoders.get(contentTopic);
         if (decoder) {
-          return decoder.decode(toProtoMessage(protoMsg));
+          return decoder.fromProtoObj(toProtoMessage(protoMsg));
         }
       }
       return Promise.resolve(undefined);