+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `Message` interface changed to ensure implementations do not omit fields.
+- `Decoder` and `Encoder` interfaces change to better express what the function members do.
 
 ## [0.29.0] - 2022-09-21